+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>io.github.dinty1</groupId>
     <artifactId>VanishTP</artifactId>
-    <version>1.3.0</version>
+    <version>2.0.0-SNAPSHOT</version>
 
     <build>
         <plugins>
